requirements.txt
SQLAlchemy
PyJWT
retry
cryptography
py
lxml
setuptools
fastapi
loguru
uvicorn
pydantic
aiohttp
pydantic_settings